Gempyp 1.0.60b0

+++ gempyp/libs/common.py
@@ -11,6 +11,11 @@ import sys
 from gempyp.libs.gem_s3_common import download_from_s3, upload_to_s3, create_s3_link
 from gempyp.config.GitLinkXML import fetchFileFromGit
 from gempyp.config import DefaultSettings
+from gempyp.engine import dataUpload
+import logging
+import requests
+from gempyp.engine.dataUpload import _getHeaders
+import re
 
 
 def read_json(file_path):
@@ -60,18 +65,21 @@ def errorHandler(logging, Error, msg="some Error Occured"):
 
 def parseMails(mail: Union[str, typing.TextIO]) -> List:
     try:
+        if(mail is not None):
+            if hasattr(mail, "read"):
+                mails = mail.read()
 
-        if hasattr(mail, "read"):
-            mails = mail.read()
-
-        elif os.path.isfile(mail):
-            file = open(mail, "r")
-            mails = file.read()
-            file.close()
-
-        mails = mail.strip()
-        mails = mails.split(",")
-        return mails
+            elif os.path.isfile(mail):
+                file = open(mail, "r")
+                mails = file.read()
+                file.close()
+            mails = mail.strip()
+            emails=[]
+            regex = r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,7}\b'
+            for mail in mails.split(","):
+                if(re.fullmatch(regex, mail)):
+                    emails.append(mail)
+            return emails
     except Exception as e:
         logging.error("Error while parsing the mails")
         logging.error(f"Error : {e}")
@@ -147,15 +155,19 @@ def moduleImports(file_name):
             return e
 
 def download_common_file(file_name,headers=None):
-    try:
+     try:
         if file_name and (file_name.__contains__('S3')):
             logging.info("File is from S3")
-            fileContent=download_from_s3(api=file_name.replace("S3:",""),username=headers.get("username",None),bridge_token=headers.get("bridge_token",None))
-            file_name = os.path.join(file_name.split(":")[-1])
-            with open(file_name, "w+") as fp:
-                fp.seek(0)
-                fp.write(fileContent)
-                fp.truncate()
+            response=download_from_s3(api=file_name.replace("S3:",""),username=headers.get("username",None),bridge_token=headers.get("bridge_token",None))
+            if(response.status_code==200):
+                file_name = os.path.join(file_name.split(":")[-1])
+                with open(file_name, "w+") as fp:
+                    fp.seek(0)
+                    fp.write(response.text)
+                    fp.truncate()
+            else:
+                logging.info(response.status_code)
+                logging.info(response.text)
         elif file_name and (file_name.__contains__('GIT')):
             logging.info("File is from GIT")
             list_url=file_name.split(":")
@@ -164,7 +176,7 @@ def download_common_file(file_name,headers=None):
             else:
                 file_name=fetchFileFromGit(list_url[2],list_url[3])
         return file_name
-    except Exception as e:
+     except Exception as e:
         traceback.print_exc()
         return e
             
@@ -198,3 +210,40 @@ def get_reason_of_failure(data, e=None):
         return exceptionarray[0]
     except:
         return e
+
+
+def validateZeroTestcases(testcaseLength):
+    if not testcaseLength:  # in case of zero testcases, we should not insert suite data 
+            logging.warning("NO TESTCASES TO RUN..... PLEASE CHECK RUN FLAGS. ABORTING.................")
+            sys.exit()
+            
+
+
+def runBaseUrls(jewel_user,base_url,username,bridgetoken):
+    if jewel_user:
+            #trying rerun of base url api in case of api failure
+            if base_url and DefaultSettings.apiSuccess == False:
+                logging.info("Retrying to call Api for getting urls")
+                DefaultSettings.getEnterPoint(base_url ,bridgetoken,username)
+            if not base_url:
+                DefaultSettings.getEnterPoint(DefaultSettings.default_baseurl ,bridgetoken, username)
+
+
+def sendMail(s_run_id,mails,bridge_token,username):
+    try:
+        # payload={"to": mails, "s_run_id": s_run_id}
+        mails["s_run_id"]=s_run_id
+        response = requests.request(
+        method="POST",
+        url=DefaultSettings.getUrls("email-api"),
+        data=json.dumps(mails),
+        headers=_getHeaders(bridge_token, username),
+        )
+        if response.status_code == 200:
+            logging.info("Report successfully sent on mail")
+        else:
+            logging.info(response.text)
+    except Exception as e:
+        traceback.print_exc()
+
+
